chore(test): split app store notification tests
