+++ tests/Feature/HandleAppStoreNotificationFeatureTest.php
@@ -1,0 +1,63 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Imdhemy\Purchases\Tests\Feature;
+
+use Illuminate\Support\Facades\Event;
+use Imdhemy\AppStore\Jws\JwsVerifier;
+use Imdhemy\Purchases\Events\AppStore\DidChangeRenewalStatus;
+use Imdhemy\Purchases\Events\AppStore\Subscribed;
+use Imdhemy\Purchases\Tests\TestCase;
+
+final class HandleAppStoreNotificationFeatureTest extends TestCase
+{
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->app->bind(JwsVerifier::class, \Imdhemy\Purchases\Tests\Doubles\JwsVerifier::class);
+    }
+
+    /** @test */
+    public function handle_app_store_server_notifications(): void
+    {
+        $this->withoutExceptionHandling();
+        Event::fake();
+        $data = json_decode(
+            file_get_contents($this->fixturesDir('appstore-server-notification.json')),
+            true,
+            512,
+            JSON_PARTIAL_OUTPUT_ON_ERROR
+        );
+        $this->post('/liap/notifications?provider=app-store', $data)->assertStatus(200);
+
+        Event::assertDispatched(DidChangeRenewalStatus::class);
+    }
+
+    /** @test */
+    public function handle_app_store_test_notification(): void
+    {
+        file_put_contents(storage_path('logs/laravel.log'), '');
+        $this->withoutExceptionHandling();
+        $signedPayload = $this->faker->appStoreTestNotification();
+
+        $this->post('/liap/notifications?provider=app-store', ['signedPayload' => $signedPayload->toString()]
+        )->assertStatus(200);
+
+        $logs = file_get_contents(storage_path('/logs/laravel.log'));
+        $this->assertStringContainsString('AppStoreV2NotificationHandler: Test notification received', $logs);
+    }
+
+    /** @test */
+    public function handle_app_store_server_notification_v2(): void
+    {
+        Event::fake();
+        $signedPayload = $this->faker->appStoreNotification();
+
+        $this->post('/liap/notifications?provider=app-store', ['signedPayload' => $signedPayload->toString()]
+        )->assertStatus(200);
+
+        Event::assertDispatched(Subscribed::class);
+    }
+}
